[TP-11536] Resolved Questions Namespace TODO
